backend/requirements/base.txt
bcrypt==3.1.3
bleach==3.3.0
boto==2.48.0
boto3==1.4.7
botocore==1.7.22
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
Django==1.11.29
django-appconf==1.0.2
django-compressor==2.1.1
django-debug-toolbar==1.8
django-environ==0.11.2
django-extensions==2.1.0
django-filter==1.0.4
django-simple-history==1.9.0
django-storages==1.7.2
djangorestframework==3.5.3
djangorestframework-expiring-authtoken==0.1.4
docutils==0.14
html5lib==0.999999999
idna==2.6
jmespath==0.9.3
olefile==0.44
openpyxl==2.5.12
pandas==0.24.2
Pillow==7.1.0
postmarker==0.11.2
psycopg2-binary==2.7.6.1
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
s3transfer==0.1.11
sendgrid==3.6.3
sendgrid-django==4.0.0
six==1.10.0
sqlparse==0.2.3
Unipath==1.1
webencodings==0.5.1
